+++ oneadapter/src/main/java/com/idanatz/oneadapter/internal/selection/OneItemDetailLookup.kt
@@ -5,7 +5,7 @@ import org.jetbrains.annotations.NotNull
 import androidx.annotation.Nullable
 import androidx.recyclerview.selection.ItemDetailsLookup
 import androidx.recyclerview.widget.RecyclerView
-import com.idanatz.oneadapter.internal.utils.toOneViewHolder
+import com.idanatz.oneadapter.internal.utils.extensions.toOneViewHolder
 
 internal class OneItemDetailLookup(private val recyclerView: RecyclerView) : ItemDetailsLookup<Long>() {