+++ CHANGELOG.md
@@ -1,5 +1,12 @@
-Change Log
+Changelog
 ==========
+
+Version 1.2.0
+----------------------------
+* Diffable is now a mandatory when working with OneAdapter
+* Bug Fixes:
+    * Paging Module not working properly with some layout manager
+    * onUnbind call now called when the view is recycled instead of detached from window
 
 Version 1.1.1
 ----------------------------